min/max fix

+++ src/lib/libcc/math/cc_math.h
@@ -6,18 +6,18 @@
 /*   By: dbutterw <dbutterw@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/17 17:21:30 by dbutterw          #+#    #+#             */
-/*   Updated: 2020/07/17 17:21:30 by dbutterw         ###   ########.fr       */
+/*   Updated: 2020/07/25 16:17:35 by dbutterw         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef CC_MATH_H
 # define CC_MATH_H
 # include <stdlib.h>
-
+# include <stdint.h>
 # undef min
 # undef max
 
-long long int 	min(long long int x, long long int y);
-long long int	max(long long int x, long long int y);
+int64_t 	min(int64_t x, int64_t y);
+int64_t		max(int64_t x, int64_t y);
 
 #endif
